requirements.txt
streamlit
langchain
PyMuPDF
faiss-cpu
requests
numpy
tiktoken
sentence-transformers
langchain-community
python-dotenv
python-telegram-bot
pydantic